requirements.txt
anyio==3.6.1
asgi-lifespan==1.0.1
asgiref==3.5.2
async-generator==1.10
asyncio==3.4.3
attrs==22.1.0
autopep8==1.7.0
bcrypt==4.0.0
beanie==1.11.9
bson==0.5.10
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.83.0
h11==0.12.0
httpcore==0.13.7
httptools==0.4.0
httpx==0.18.2
idna==3.3
iniconfig==1.1.1
JSON-log-formatter==0.5.1
makefun==1.15.0
motor==3.0.0
multidict==6.0.2
outcome==1.2.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.1.3
pytest-asyncio==0.19.0
pytest-env==0.6.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.17.0
websockets==10.3
wrapt==1.14.1
yarl==1.8.1
